+++ app/src/main/java/com/dede/android_eggs/inject/EasterEggModules.kt
@@ -9,6 +9,7 @@ import com.android_k.egg.AndroidKitKatEasterEgg
 import com.android_l.egg.AndroidLollipopEasterEgg
 import com.android_m.egg.AndroidMarshmallowEasterEgg
 import com.android_n.egg.AndroidNougatEasterEgg
+import com.android_next.egg.AndroidNextEasterEgg
 import com.android_o.egg.AndroidOreoEasterEgg
 import com.android_p.egg.AndroidPieEasterEgg
 import com.android_q.egg.AndroidQEasterEgg
@@ -31,7 +32,7 @@ import javax.inject.Singleton
 
 @Module(
     includes = [
-//        AndroidNextEasterEgg::class,
+        AndroidNextEasterEgg::class,
         AndroidVEasterEgg::class,
         AndroidUEasterEgg::class,
         AndroidTEasterEgg::class,