+++ app/src/main/java/com/dede/android_eggs/views/settings/compose/prefs/IconVisualEffectsPref.kt
@@ -1,6 +1,7 @@
 package com.dede.android_eggs.views.settings.compose.prefs
 
 import android.content.Context
+import android.os.Build
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Animation
 import androidx.compose.runtime.Composable
@@ -17,8 +18,12 @@ object IconVisualEffectsPrefUtil {
     const val ACTION_CHANGED = "com.dede.android_eggs.IconVisualEffectsChanged"
     const val KEY_ICON_VISUAL_EFFECTS = "pref_key_icon_visual_effects"
 
+    fun isSupported(): Boolean {
+        return Build.VERSION.SDK_INT >= Build.VERSION_CODES.O
+    }
+
     fun isEnable(context: Context): Boolean {
-        return SettingPrefUtil.getValue(
+        return isSupported() && SettingPrefUtil.getValue(
             context, KEY_ICON_VISUAL_EFFECTS, SettingPrefUtil.OFF
         ) == SettingPrefUtil.ON
     }